Condense route logic
Create utility functions to be used across neighborhood and listing route selectors for:
- Memoize route logic
- Create unique segment logic
- Get segment and stop positions logic
