+++ taui/src/selectors/network-neighborhood-routes.js
@@ -1,37 +1,12 @@
 // @flow
 import lonlat from '@conveyal/lonlat'
 import get from 'lodash/get'
-import memoize from 'lodash/memoize'
 import {createSelector} from 'reselect'
 
-import createTransitiveRoutes from '../utils/create-transitive-routes'
-import createDirectRoutes from '../utils/create-direct-routes'
+import {memoizedTransitiveRoutes} from '../utils/memoize-routes'
+import uniqueSegments from '../utils/make-unique-segments'
 
 import selectActiveNetworkIndex from './active-network-index'
-
-/**
- * This assumes loaded query, paths, and targets.
- */
-const memoizedTransitiveRoutes = memoize(
-  (n, i, s, e, car) => (car ? createDirectRoutes(n, s, e) : createTransitiveRoutes(n, s, e)),
-  (n, i, s, e, car) =>
-    `${n.name}-${i}-${n.originPoint.x}-${n.originPoint.y}-${lonlat.toString(e.position)}-car-${car}`
-)
-
-const routeToString = s =>
-  s.map(s => `${s.name}-${s.backgroundColor}-${s.type}`).join('-')
-
-const uniqueSegments = routeSegments => {
-  const foundKeys = {}
-  return (routeSegments || []).reduce((uniqueRoutes, route) => {
-    const key = routeToString(route)
-    if (!foundKeys[key]) {
-      foundKeys[key] = true
-      return [...uniqueRoutes, route]
-    }
-    return uniqueRoutes
-  }, [])
-}
 
 export default createSelector(
   selectActiveNetworkIndex,
